+++ packages/frontend/src/components/providers/game/index.jsx
@@ -1,52 +1,21 @@
 import React, { useReducer, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { GameContext } from 'contexts/game';
+import { addTiles } from 'contexts/game/actions';
 import { baseState } from 'contexts';
 import { gameReducer } from 'contexts/game/reducer';
-import { addTiles } from 'contexts/game/actions';
 import { mockPhotos } from '__mocks__/api/mockPhotos';
-
-const shufflePhotos = (photos) => {
-  let currentIndex = photos.length,
-    randomIndex;
-
-  while (currentIndex > 0) {
-    randomIndex = Math.floor(Math.random() * currentIndex);
-    currentIndex--;
-
-    [photos[currentIndex], photos[randomIndex]] = [
-      photos[randomIndex],
-      photos[currentIndex],
-    ];
-  }
-  return photos;
-};
+import { createTiles } from 'helpers/createTiles';
 
 const GameProvider = ({ children, providedState = null } = {}) => {
   const initialState = providedState || baseState.game;
   const [state, dispatch] = useReducer(gameReducer, initialState);
 
-  //should we eventually change this use effect to a useMemo?
+  const gameTiles = createTiles(mockPhotos);
   useEffect(() => {
-    // change this to ladash duplicate *see note
-    const unshuffledPhotos = [...mockPhotos].concat(mockPhotos);
-
-    const tilePhotos = shufflePhotos(unshuffledPhotos);
-
-    const tiles = [];
-
-    tilePhotos.forEach((photo) => {
-      tiles.push({
-        isMatched: false,
-        faceUp: false,
-        photo: photo,
-      });
-      return tiles;
-    });
-
-    dispatch(addTiles(tiles));
+    dispatch(addTiles(gameTiles));
   }, []);
-  console.log(state, 'state tiles');
+
   return (
     <GameContext.Provider value={{ state, dispatch }}>
       {children}
